backend_app/requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiohttp_socks==0.9.1
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asttokens==3.0.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
Brotli==1.1.0
browser-cookie3==0.19.1
cachetools==5.5.0
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
cryptography==44.0.0
cssselect2==0.7.0
curl_cffi==0.7.4
cycler==0.12.1
debugpy==1.8.9
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
distro==1.9.0
dnspython==2.7.0
duckduckgo_search==6.3.7
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.6
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.5.0
fuzzywuzzy==0.18.0
g4f==0.3.9.7
gast==0.4.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-genai==0.3.0
google-generativeai==0.8.3
google-pasta==0.2.0
googleapis-common-protos==1.66.0
grpcio==1.68.1
grpcio-status==1.68.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.4
jiter==0.8.0
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==3.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.4
matplotlib-inline==0.1.7
mdurl==0.1.2
ml-dtypes==0.4.1
mss==10.0.0
multidict==6.1.0
mysql-connector==2.2.9
namex==0.0.8
nest-asyncio==1.6.0
nodriver==0.38.post1
numpy==2.0.2
oauthlib==3.2.2
openai==1.55.3
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.0.0
platformdirs==4.3.6
primp==0.8.1
prompt_toolkit==3.0.48
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
psutil==6.1.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pycryptodomex==3.21.0
pydantic==2.10.2
pydantic_core==2.27.1
Pygments==2.18.0
pymongo==4.10.1
PyMySQL==1.1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
python-socks==2.5.3
pytz==2024.2
pyzmq==26.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
scikit-learn==1.6.1
scipy==1.13.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.41.3
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.18.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
threadpoolctl==3.5.0
tinycss2==1.4.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.32.1
wcwidth==0.2.13
webencodings==0.5.1
websockets==13.1
Werkzeug==3.1.3
wrapt==1.17.0
yarl==1.18.3
zipp==3.21.0
